requirements.txt
absl-py==1.0.0
ansicon==1.89.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.3.4
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.81
bleach==5.0.1
blessed==1.19.1
cachetools==4.2.4
catboost==1.2.5
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
click==8.1.3
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.3
edge==0.4.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
featurewiz==0.5.7
filelock==3.16.1
Flask==2.2.2
flatbuffers==2.0
fonttools==4.38.0
fqdn==1.5.1
fsspec==2024.9.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.43.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.6.0
httplib2==0.20.4
huggingface-hub==0.25.1
idna==3.3
imageio==2.22.4
imbalanced-learn==0.12.2
imblearn==0.0
importlib_metadata==7.1.0
inquirer==3.0.0
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter-http-over-ws==0.0.8
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.3
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.25.2
oauthlib==3.1.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.6.0
plotly==5.10.0
pomegranate==0.14.8
prometheus-client==0.15.0
prompt-tool-kit==1.0.14
prompt-toolkit==3.0.36
protobuf==3.19.1
psutil==5.9.4
pure-eval==0.2.2
PyArabic==0.6.15
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
PyPDF2==3.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-editor==1.0.4
python-inquirer==1.0.3
python-json-logger==2.0.4
pytz==2021.1
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==24.0.1
readchar==4.0.3
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.8
safetensors==0.4.5
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data==0.6.2
tenacity==8.0.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tokenizers==0.20.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
transformers==4.45.2
typing_extensions==4.0.1
tzdata==2024.1
uri-template==1.2.0
urllib3==1.26.8
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.13.3
xgboost==2.0.3
xlrd==2.0.1
zipp==3.7.0
